+++ src/utils.rs
@@ -218,6 +218,15 @@ where
 {
 }
 
+unsafe impl<T, T2, const N: usize, const N2: usize, A: Alignment, A2: Alignment>
+    Specialize<Option<T2>, N, N2, A, A2> for Option<T>
+where
+    T: Specialize<T2, N, N2, A, A2>,
+    Length<N>: SupportedLength,
+    Length<N2>: SupportedLength,
+{
+}
+
 unsafe impl<R, R2, const N: usize, const N2: usize, A: Alignment, A2: Alignment>
     Specialize<fn() -> R2, N, N2, A, A2> for fn() -> R
 where